+++ src/cpu-template-helper/src/strip/aarch64.rs
@@ -1,8 +1,8 @@
 // Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use vmm::guest_config::templates::aarch64::RegisterModifier;
-use vmm::guest_config::templates::CustomCpuTemplate;
+use vmm::cpu_config::aarch64::custom_cpu_template::RegisterModifier;
+use vmm::cpu_config::templates::CustomCpuTemplate;
 
 use crate::strip::strip_common;
 use crate::utils::aarch64::RegModifierMap;
@@ -30,8 +30,8 @@ pub fn strip(templates: Vec<CustomCpuTemplate>) -> Vec<CustomCpuTemplate> {
 
 #[cfg(test)]
 mod tests {
-    use vmm::guest_config::templates::aarch64::RegisterModifier;
-    use vmm::guest_config::templates::RegisterValueFilter;
+    use vmm::cpu_config::aarch64::custom_cpu_template::RegisterModifier;
+    use vmm::cpu_config::templates::RegisterValueFilter;
 
     use super::*;
     use crate::utils::aarch64::reg_modifier;
