+++ src/cpu-template-helper/src/utils/x86_64.rs
@@ -4,11 +4,11 @@
 use std::collections::HashMap;
 use std::fmt::Display;
 
-use vmm::guest_config::templates::x86_64::{
+use vmm::cpu_config::templates::RegisterValueFilter;
+use vmm::cpu_config::x86_64::cpuid::KvmCpuidFlags;
+use vmm::cpu_config::x86_64::custom_cpu_template::{
     CpuidLeafModifier, CpuidRegister, CpuidRegisterModifier, RegisterModifier,
 };
-use vmm::guest_config::templates::RegisterValueFilter;
-use vmm::guest_config::x86_64::cpuid::KvmCpuidFlags;
 
 use super::{ModifierMapKey, ModifierMapValue};
 
@@ -223,8 +223,8 @@ pub(crate) use {cpuid_leaf_modifier, cpuid_reg_modifier, msr_modifier};
 
 #[cfg(test)]
 mod tests {
-    use vmm::guest_config::templates::x86_64::CpuidRegister::*;
-    use vmm::guest_config::templates::x86_64::CpuidRegisterModifier;
+    use vmm::cpu_config::x86_64::custom_cpu_template::CpuidRegister::*;
+    use vmm::cpu_config::x86_64::custom_cpu_template::CpuidRegisterModifier;
 
     use super::*;
     use crate::utils::x86_64::{cpuid_leaf_modifier, cpuid_reg_modifier, msr_modifier};
