+++ src/cpu-template-helper/src/verify/x86_64.rs
@@ -1,7 +1,7 @@
 // Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use vmm::guest_config::templates::CustomCpuTemplate;
+use vmm::cpu_config::templates::CustomCpuTemplate;
 
 use super::{verify_common, Error};
 use crate::utils::x86_64::{CpuidModifierMap, MsrModifierMap};
@@ -22,12 +22,12 @@ pub fn verify(cpu_template: CustomCpuTemplate, cpu_config: CustomCpuTemplate) ->
 mod tests {
     use std::collections::HashMap;
 
-    use vmm::guest_config::templates::x86_64::CpuidRegister::*;
-    use vmm::guest_config::templates::x86_64::{
+    use vmm::cpu_config::templates::RegisterValueFilter;
+    use vmm::cpu_config::x86_64::cpuid::KvmCpuidFlags;
+    use vmm::cpu_config::x86_64::custom_cpu_template::CpuidRegister::*;
+    use vmm::cpu_config::x86_64::custom_cpu_template::{
         CpuidLeafModifier, CpuidRegisterModifier, RegisterModifier,
     };
-    use vmm::guest_config::templates::RegisterValueFilter;
-    use vmm::guest_config::x86_64::cpuid::KvmCpuidFlags;
 
     use super::*;
     use crate::utils::x86_64::{
