+++ src/vmm/src/cpu_config/aarch64/mod.rs
@@ -1,10 +1,12 @@
 // Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-/// Module with CPU templates for aarch64
+/// Module for custom CPU templates
+pub mod custom_cpu_template;
+/// Module for static CPU templates
 pub mod static_cpu_templates;
-
-pub use static_cpu_templates::*;
+/// Module with test utils for custom CPU templates
+pub mod test_utils;
 
 use super::templates::CustomCpuTemplate;
 use crate::arch::regs::{Aarch64Register, Error as ArchError};
