+++ src/vmm/src/cpu_config/aarch64/static_cpu_templates/mod.rs
