+++ src/vmm/src/cpu_config/aarch64/test_utils.rs
@@ -1,12 +1,10 @@
 // Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-#[cfg(target_arch = "aarch64")]
-use crate::guest_config::templates::aarch64::RegisterModifier;
-use crate::guest_config::templates::{CustomCpuTemplate, RegisterValueFilter};
+use crate::cpu_config::aarch64::custom_cpu_template::RegisterModifier;
+use crate::cpu_config::templates::{CustomCpuTemplate, RegisterValueFilter};
 
 /// Test CPU template in JSON format
-#[cfg(target_arch = "aarch64")]
 pub const TEST_TEMPLATE_JSON: &str = r#"{
     "reg_modifiers":  [
         {
@@ -23,7 +21,6 @@ pub const TEST_TEMPLATE_JSON: &str = r#"{
 /// Test CPU template in JSON format but has an invalid field for the architecture.
 /// "msr_modifiers" is the field name for the model specific registers for
 /// defined by x86 CPUs.
-#[cfg(target_arch = "aarch64")]
 pub const TEST_INVALID_TEMPLATE_JSON: &str = r#"{
     "msr_modifiers":  [
         {
@@ -34,7 +31,6 @@ pub const TEST_INVALID_TEMPLATE_JSON: &str = r#"{
 }"#;
 
 /// Builds a sample custom CPU template
-#[cfg(target_arch = "aarch64")]
 pub fn build_test_template() -> CustomCpuTemplate {
     CustomCpuTemplate {
         reg_modifiers: vec![
