+++ src/vmm/src/cpu_config/mod.rs
@@ -1,9 +1,9 @@
 // Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-/// Module for types used for custom CPU templates
+/// Module with types used for custom CPU templates
 pub mod templates;
-/// Module for types used for custom CPU templates
+/// Module with ser/de utils for custom CPU templates
 pub mod templates_serde;
 
 /// Module containing type implementations needed for x86 CPU configuration
