+++ src/vmm/src/cpu_config/templates.rs
@@ -1,29 +1,29 @@
 // Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-/// Module with cpu templates for x86_64
-#[cfg(target_arch = "x86_64")]
-pub mod x86_64;
-
-/// Utility module for testing guest_configuration.
-pub mod test_utils;
+// /// Module with cpu templates for x86_64
+// #[cfg(target_arch = "x86_64")]
+// pub mod x86_64;
+//
+// /// Utility module for testing cpu_configuration.
+// pub mod test_utils;
 
 #[cfg(target_arch = "x86_64")]
 mod common_types {
-    pub use crate::guest_config::templates::x86_64::CustomCpuTemplate;
-    pub use crate::guest_config::x86_64::static_cpu_templates::StaticCpuTemplate;
-    pub use crate::guest_config::x86_64::{CpuConfiguration, Error as GuestConfigError};
+    pub use crate::cpu_config::x86_64::custom_cpu_template::CustomCpuTemplate;
+    pub use crate::cpu_config::x86_64::static_cpu_templates::StaticCpuTemplate;
+    pub use crate::cpu_config::x86_64::{test_utils, CpuConfiguration, Error as GuestConfigError};
 }
 
 /// Module with cpu templates for aarch64
-#[cfg(target_arch = "aarch64")]
-pub mod aarch64;
+// #[cfg(target_arch = "aarch64")]
+// pub mod aarch64;
 
 #[cfg(target_arch = "aarch64")]
 mod common_types {
-    pub use crate::guest_config::aarch64::static_cpu_templates::StaticCpuTemplate;
-    pub use crate::guest_config::aarch64::{CpuConfiguration, Error as GuestConfigError};
-    pub use crate::guest_config::templates::aarch64::CustomCpuTemplate;
+    pub use crate::cpu_config::aarch64::custom_cpu_template::CustomCpuTemplate;
+    pub use crate::cpu_config::aarch64::static_cpu_templates::StaticCpuTemplate;
+    pub use crate::cpu_config::aarch64::{test_utils, CpuConfiguration, Error as GuestConfigError};
 }
 
 use std::borrow::Cow;
@@ -40,7 +40,7 @@ pub enum GetCpuTemplateError {
     #[cfg(target_arch = "x86_64")]
     /// Failed to get CPU vendor information.
     #[error("Failed to get CPU vendor information: {0}")]
-    GetCpuVendor(crate::guest_config::x86_64::cpuid::common::GetCpuidError),
+    GetCpuVendor(crate::cpu_config::x86_64::cpuid::common::GetCpuidError),
     /// CPU Vendor mismatched between the actual CPU and CPU template.
     #[error("CPU vendor mismatched between actual CPU and CPU template.")]
     CpuVendorMismatched,
