+++ src/vmm/src/cpu_config/templates_serde.rs
