+++ src/vmm/src/cpu_config/x86_64/cpuid/amd/normalize.rs
@@ -1,11 +1,11 @@
 // Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::guest_config::x86_64::cpuid::common::{get_vendor_id_from_host, GetCpuidError};
-use crate::guest_config::x86_64::cpuid::normalize::{
+use crate::cpu_config::x86_64::cpuid::common::{get_vendor_id_from_host, GetCpuidError};
+use crate::cpu_config::x86_64::cpuid::normalize::{
     get_range, set_bit, set_range, CheckedAssignError,
 };
-use crate::guest_config::x86_64::cpuid::{
+use crate::cpu_config::x86_64::cpuid::{
     cpuid, cpuid_count, CpuidEntry, CpuidKey, CpuidRegisters, CpuidTrait, KvmCpuidFlags,
     MissingBrandStringLeaves, BRAND_STRING_LENGTH, VENDOR_ID_AMD,
 };
@@ -436,7 +436,7 @@ mod tests {
     use std::collections::BTreeMap;
 
     use super::*;
-    use crate::guest_config::x86_64::cpuid::AmdCpuid;
+    use crate::cpu_config::x86_64::cpuid::AmdCpuid;
 
     #[test]
     fn test_update_structured_extended_entry_invalid() {
