+++ src/vmm/src/cpu_config/x86_64/static_cpu_templates/mod.rs
